fix: cast zipped data to Uint8Array before creating Blob to ensure correct type

+++ src/workers/zipWorker.ts
@@ -229,9 +229,9 @@ onmessage = async (evt: MessageEvent<WorkerIn>) => {
       postProgress({ type: 'overall', processed, total: names.length });
     }
 
-  const zipped = zipSync(out, { level: 6 });
-  const result = new Blob([zipped], { type: 'application/zip' });
-  (postMessage as any)({ type: 'done', jobId, blob: result } satisfies WorkerOut);
+  const zipped = zipSync(out, { level: 6 }) as Uint8Array;
+const result = new Blob([zipped], { type: 'application/zip' });
+(postMessage as any)({ type: 'done', jobId, blob: result } satisfies WorkerOut);
   } catch (err) {
     (postMessage as any)({ type: 'progress', payload: { type: 'error', name: '(worker)', reason: (err as Error).message } });
   }
